feature: Motif FInder
Initialized motif finder program program and necessary files, a bash shell script tool and a README markdown.

Software intended to accept a text file containing one or more motifs to search through a target .fasta file and count the number of occurances of each motif and extract the genes with one or more of the target motifs.